requirements.txt
beautifulsoup4>=4.12
python-dotenv>=1.0
fastapi>=0.100
lxml>=5.0
playwright==1.50.0
uvicorn>=0.20
